+++ internal/acserver/tcp_message_split.go
@@ -49,12 +49,14 @@ func (t SectorSplitMessageHandler) OnMessage(conn net.Conn, p *Packet) error {
 	t.state.BroadcastOthersTCP(bw, entrant.CarID)
 
 	go func() {
-		err := t.plugin.OnSectorCompleted(Split{
+		// @TODO I think this is the first time we get told about cuts
+		// @TODO if we decide to do ballast for cut this might be the best place
+		err := t.state.CompleteSector(Split{
 			Car:   *entrant,
 			Index: splitIndex,
 			Time:  splitTime,
 			Cuts:  cuts,
-		})
+		}, entrant)
 
 		if err != nil {
 			t.logger.WithError(err).Error("On sector completed plugin returned an error")